Move Field to file
