+++ src/scippnexus/v2/application_definitions/nxcansas/nxcansas.py
@@ -6,7 +6,8 @@ from typing import Any, Callable, Dict, Literal, Optional, Union
 import scipp as sc
 
 from ....typing import H5Group
-from ...base import Field, Group, NXobject, base_definitions, create_field
+from ...base import Group, NXobject, base_definitions, create_field
+from ...field import Field
 from ...nxdata import NXdata
